requirements.txt
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.0.3
fonttools==4.56.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
narwhals==1.28.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
open3d==0.19.0
packaging==24.2
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==6.0.0
plyfile==1.0.3
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.1
PyWavelets==1.4.1
pywin32==308
pyzmq==26.2.1
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rpds-py==0.20.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.5.0
tifffile==2023.7.10
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.6.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
zipp==3.20.2
